Abstracted pickRandomPiece() logic into a tetromino class that is randomized automatically.
